requirements.txt
flask
celery
anti_useragent
mongoengine 
elasticsearch_dsl  
elasticsearch
flask_sqlalchemy
pymongo
pymysql
selenium
kafka-python